feat: Add scribble tool and bump version
This commit introduces a new tool called `scribble`.

The `scribble` tool allows creating or replacing a file on disk with the given content. If the file is a `.go` file, it will be checked for errors, formatted, and have its imports fixed.

This commit also bumps the minor version to 0.3.0.